pyproject.toml
[project]
name = "hrmis"
version = "0.1.0"
description = "HR Management Information System for Sindh Government"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "xlrd>=2.0.2",
    "xlsxwriter>=3.2.9",
]

[dependency-groups]
dev = [
    "click>=8.1.0",
    "odoo-stubs>=0.1.dev0",
    "pyright>=1.1.407",
    "ruff>=0.14.6",
]

# Pyright Config
[tool.pyright]
include = ["modules"]
exclude = ["**/__pycache__"]
reportMissingImports = "warning"
reportMissingTypeStubs = false

# Ruff Config
[tool.ruff]
line-length = 88
target-version = "py312"
